somth new 3

+++ src/main/java/just/fun/chess/TrainingExample.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class TrainingExample {
 
-    private final float input[];
-    private final float score[];
+    private final byte input[];
+    private final byte score[];
 }
